new name vpnping.sh

+++ vpnping.sh
@@ -3,6 +3,8 @@
 # **************************************************************************
 # * (C)opyright 2017         by Ruben Carlo Benante                        *
 # *                                                                        *
+# * VPNPing.sh ping your VPN servers and find the fastest                  *
+# *                                                                        *
 # * This program is free software; you can redistribute it and/or modify   *
 # *  it under the terms of the GNU General Public License as published by  *
 # *  the Free Software Foundation version 2 of the License.                *
@@ -28,18 +30,30 @@ VERSION="20171118.225122"
 Help()
 {
     cat << EOF
-    nordping - Ping a bunch of NordVPN servers to test the fastest
-    Usage: $ ./nordping.sh [-v] ( [-h|-V] | [-p TCP|UDP ] )
+    vpnping - Ping a bunch of VPNs ovpn file (such as SurfShark VPN or NordVPN) servers to test the fastest ping response
+    Usage: $ ./vpnping.sh [-v] ( [-h|-V] | [-p TCP|UDP ] )
 
     Options:
       -h, --help       Show this help.
       -V, --version    Show version.
       -v, --verbose    Turn verbose mode on (cumulative).
-      -p, --protocol   Set the protocol to ping: TCP (default) or UDP
+      -p, --protocol   Set the protocol to ping: tcp (default) or udp (lower case)
       -d, --directory  Sets the directory where to find all *.ovpn files (defaults to ./ovpn-files)
       -f, --filter     Filter by prefix
-      -w, --wget       Download ovpn files from nordvpn site to directory ./ovpn-files
-      -e, --stderr     Print also to stderr. Useful if piping, e.g.: $ ./nordping.sh | sort -nk3
+      -w, --wget       Download ovpn files from nordvpn site to directory ./ovpn-nord or surfshark to ./ovpn-shark
+      -e, --stderr     Print also to stderr. Useful if piping, e.g.: $ ./vpnping.sh | sort -nk3
+      -s, --server     Currently available options: -s shark or -s nord
+
+    Example:
+        Example of a command: 
+            ~/dir/$ ./vpnping.sh -f sp -p udp -d ./dir/ovpn-shark
+
+        Example of an answer:
+            ./ovpn-shark/sp1.sharkvpn.com.udp1122.ovpn         189.187.142.116          3000.33         error
+            ./ovpn-shark/sp2.sharkvpn.com.udp1122.ovpn         175.183.176.2   418.501
+
+
+
     Exit status:
        0, if ok.
        1, some error occurred.
@@ -50,6 +64,7 @@ Help()
     Author:
             Written by Ruben Carlo Benante <rcb@beco.cc>
             Created: 2017-02-18
+            Major mod: 2021-08-13
 EOF
     exit 1
 }
@@ -57,9 +72,9 @@ EOF
 # The copyright function
 Copyr()
 {
-    echo "nordping - ${VERSION}"
+    echo "vpnping - ${VERSION}"
     echo
-    echo 'Copyright (C) 2017 Ruben Carlo Benante <rcb@beco.cc>, GNU GPL version 2'
+    echo 'Copyright (C) 2017-2021+ Ruben Carlo Benante <rcb@beco.cc>, GNU GPL version 2'
     echo '<http://gnu.org/licenses/gpl.html>. This  is  free  software:  you are free to change and'
     echo 'redistribute it. There is NO WARRANTY, to the extent permitted by law. USE IT AS IT IS. The author'
     echo 'takes no responsability to any damage this software may inflige in your data.'
@@ -70,43 +85,46 @@ Copyr()
 # The download function
 Download()
 {
-    echo "Starting nordping.sh script, by beco, version ${VERSION}..."
+    echo "Starting vpnping.sh script, by beco, version ${VERSION}..."
     echo "Verbose level: $verbose"
     date
     TRUST="no"
-    wget -q --show-progress https://nordvpn.com/api/files/zip -O config.zip || { echo "Can not download ovpn files" ; exit 1 ; }
-    md5sum -c --status config.zip.md5
-    SQ=$?
-    if (( "10#0${SQ}" != 0 )) ; then
-        echo "Cannot verify the md5sum signature. File corrupted or changed."
-        echo "Type 'yes' only if you are sure in the next 15 seconds."
-        read -t 15 -p "Do you trust the new file? " TRUST
-        echo
-    else
-        echo "md5sum signature checked OK"
-        TRUST="sureOK"
-    fi
-    if [[ "${TRUST}" == "yes" || "${TRUST}" == "sureOK" ]] ; then
-        unzip config.zip -d ovpn-files
-        echo "Downloaded and unzipped successfuly"
-        if [[ "${TRUST}" != "sureOK" ]] ; then
+
+    if [ "$VPNSERVER" == "nord" ] ; then
+        wget -q --show-progress https://nordvpn.com/api/files/zip -O config.zip || { echo "Can not download ovpn files" ; exit 1 ; }
+        md5sum -c --status config.zip.md5
+        SQ=$?
+        if (( "10#0${SQ}" != 0 )) ; then
+            echo "Cannot verify the md5sum signature. File corrupted or changed."
+            echo "Type 'yes' only if you are sure in the next 15 seconds."
+            read -t 15 -p "Do you trust the new file? " TRUST
             echo
-            echo "Warning: using a non-validated md5sum file"
-            echo "Please report an issue (subject new md5sum) at https://github.com/drbeco/nordping"
-            echo "New md5sum:"
-            md5sum config.zip
-            echo "Old mdrsum:"
-            cat config.zip.md5 | grep config.zip
-            echo
+        else
+            echo "md5sum signature checked OK"
+            TRUST="sureOK"
+        fi
+        if [[ "${TRUST}" == "yes" || "${TRUST}" == "sureOK" ]] ; then
+            unzip config.zip -d ovpn-$VPNSERVER
+            echo "Downloaded and unzipped successfuly"
+            if [[ "${TRUST}" != "sureOK" ]] ; then
+                echo
+                echo "Warning: using a non-validated md5sum file"
+                echo "Please report an issue (subject new md5sum) at https://github.com/drbeco/vpnping"
+                echo "New md5sum:"
+                md5sum config.zip
+                echo "Old mdrsum:"
+                cat config.zip.md5 | grep config.zip
+                echo
+                rm -rf config.zip
+                exit 1
+            fi
             rm -rf config.zip
+            exit 0
+        else
+            rm -rf config.zip
+            echo "Download/installation aborted"
             exit 1
         fi
-        rm -rf config.zip
-        exit 0
-    else
-        rm -rf config.zip
-        echo "Download/installation aborted"
-        exit 1
     fi
 }
 
@@ -117,8 +135,10 @@ main()
     DIR="./ovpn-files"
     PREFIX=""
     PRERR=0
+    VPNSERVER="shark"
+    QNPROTO="tcp"
     #getopt example with switch/case
-    while getopts "hVvp:d:f:we" FLAG; do
+    while getopts "hVvp:d:f:wes:" FLAG; do
         case $FLAG in
             h)
                 Help
@@ -138,6 +158,9 @@ main()
             f)
                 PREFIX=$OPTARG
                 ;;
+            s)
+                VPNSERVER=$OPTARG
+                ;;
             w)
                 Download
                 ;;
@@ -151,17 +174,21 @@ main()
     done
 
     if [ "$verbose" -gt "0" ] ; then
-        echo "Starting nordping.sh script, by beco, version ${VERSION}"...
+        echo "Starting vpnping.sh script, by beco, version ${VERSION}"...
         echo "Verbose level: $verbose"
         date
     fi
 
-    if [ "$QNPROTO" == "UDP" ]; then
+    if [ "$QNPROTO" == "udp" ]; then
         QPROTO="1194"
     else
         QPROTO="443"
     fi
 
+    if [ "$verbose" -gt "1" ] ; then
+        echo "Using $VPNSERVER VPN server"
+        DIR="./ovpn-$VPNSERVER"
+    fi
     if [ "$verbose" -gt "0" ]; then
         echo "Protocol: $QPROTO"
     fi
@@ -169,7 +196,7 @@ main()
         echo "ovpn file                                          protocol           average  timeout"
     fi
 
-    for F in "$DIR"/"$PREFIX"*.ovpn ; do
+    for F in "$DIR"/"$PREFIX"*"$QNPROTO"*.ovpn ; do
         [ ! -e $F ] && echo -e 'Please give a valid path to where you downloaded ovpn files.\n(Try "-w" to automatically download them first or "-h" for help.)' && break
         LIN=`grep "remote " $F`
         PROTO=`echo $LIN | cut -d' ' -f3`
